+++ infra/helm.tf
@@ -1,0 +1,36 @@
+provider "kubernetes" {
+  host                   = module.eks.cluster_endpoint
+  cluster_ca_certificate = base64decode(module.eks.cluster_certificate_authority_data)
+  exec {
+    api_version = "client.authentication.k8s.io/v1beta1"
+    command     = "aws"
+    args        = ["eks", "get-token", "--cluster-name", module.eks.cluster_name]
+  }
+}
+
+provider "helm" {
+  kubernetes {
+    host                   = module.eks.cluster_endpoint
+    cluster_ca_certificate = base64decode(module.eks.cluster_certificate_authority_data)
+    exec {
+      api_version = "client.authentication.k8s.io/v1beta1"
+      command     = "aws"
+      args        = ["eks", "get-token", "--cluster-name", module.eks.cluster_name]
+    }
+  }
+}
+
+resource "helm_release" "aws_load_balancer_controller" {
+  name       = "aws-load-balancer-controller"
+  repository = "https://aws.github.io/eks-charts"
+  chart      = "aws-load-balancer-controller"
+  version    = "1.10.0"
+  namespace  = "kube-system"
+
+  set { name = "clusterName";                                                     value = module.eks.cluster_name }
+  set { name = "serviceAccount.create";                                           value = "true" }
+  set { name = "serviceAccount.name";                                             value = "aws-load-balancer-controller" }
+  set { name = "serviceAccount.annotations.eks\\.amazonaws\\.com/role-arn";      value = aws_iam_role.alb_controller.arn }
+
+  depends_on = [module.eks]
+}